+++ src/Contracts/ContainerInterface.php
@@ -19,7 +19,6 @@ namespace CloudCreativity\LaravelJsonApi\Contracts;
 
 use CloudCreativity\LaravelJsonApi\Contracts\Adapter\ResourceAdapterInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Auth\AuthorizerInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Auth\ResourceAuthorizerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validators\ValidatorProviderInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as BaseContainerInterface;
 
@@ -30,14 +29,6 @@ use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as BaseContainerInterfac
  */
 interface ContainerInterface extends BaseContainerInterface
 {
-
-    /**
-     * Get the domain record type for the supplied JSON API resource type.
-     *
-     * @param string $resourceType
-     * @return string
-     */
-    public function getType($resourceType);
 
     /**
      * Get a resource adapter for a domain record.
@@ -95,7 +86,7 @@ interface ContainerInterface extends BaseContainerInterface
      * Get a resource authorizer by domain record.
      *
      * @param $record
-     * @return ResourceAuthorizerInterface|null
+     * @return AuthorizerInterface|null
      *      the authorizer, if there is one.
      */
     public function getAuthorizer($record);
@@ -104,7 +95,7 @@ interface ContainerInterface extends BaseContainerInterface
      * Get a resource authorizer by domain record type.
      *
      * @param $type
-     * @return ResourceAuthorizerInterface|null
+     * @return AuthorizerInterface|null
      *      the authorizer, if there is one.
      */
     public function getAuthorizerByType($type);
@@ -113,13 +104,13 @@ interface ContainerInterface extends BaseContainerInterface
      * Get a resource authorizer by JSON API type.
      *
      * @param $resourceType
-     * @return ResourceAuthorizerInterface|null
+     * @return AuthorizerInterface|null
      *      the authorizer, if there is one.
      */
     public function getAuthorizerByResourceType($resourceType);
 
     /**
-     * Get a re-usable authorizer by name.
+     * Get a multi-resource authorizer by name.
      *
      * @param string $name
      * @return AuthorizerInterface