+++ src/Contracts/Http/Requests/RequestInterface.php
@@ -30,6 +30,13 @@ interface RequestInterface
 {
 
     /**
+     * Get the domain record type that is subject of the request.
+     *
+     * @return string
+     */
+    public function getType();
+
+    /**
      * What resource type does the request relate to?
      *
      * @return string|null