+++ tests/lib/Unit/ContainerTest.php
@@ -20,11 +20,11 @@ namespace CloudCreativity\LaravelJsonApi\Tests\Unit;
 use CloudCreativity\LaravelJsonApi\Container;
 use CloudCreativity\LaravelJsonApi\Contracts\Adapter\ResourceAdapterInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Auth\AuthorizerInterface;
-use CloudCreativity\LaravelJsonApi\Contracts\Auth\ResourceAuthorizerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Resolver\ResolverInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Validators\ValidatorProviderInterface;
 use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use Illuminate\Container\Container as IlluminateContainer;
+use League\Flysystem\AdapterInterface;
 use Neomerx\JsonApi\Contracts\Schema\SchemaProviderInterface;
 
 class ContainerTest extends TestCase
@@ -216,7 +216,7 @@ class ContainerTest extends TestCase
 
     public function testResourceAuthorizer()
     {
-        $authorizer = $this->createMock(ResourceAuthorizerInterface::class);
+        $authorizer = $this->createMock(AuthorizerInterface::class);
 
         $this->illuminateContainer
             ->expects($this->once())
@@ -236,8 +236,8 @@ class ContainerTest extends TestCase
 
     public function testResourceAuthorizerCreateReturnsNull()
     {
-        $this->resolver->method('getAuthorizerByResourceType')->willReturn(ResourceAuthorizerInterface::class);
-        $this->illuminateContainer->expects($this->never())->method('make')->with(ResourceAuthorizerInterface::class);
+        $this->resolver->method('getAuthorizerByResourceType')->willReturn(AuthorizerInterface::class);
+        $this->illuminateContainer->expects($this->never())->method('make')->with(AuthorizerInterface::class);
 
         $this->assertNull($this->container->getAuthorizerByResourceType('posts'));
         $this->assertNull($this->container->getAuthorizerByType(\stdClass::class));
@@ -252,7 +252,7 @@ class ContainerTest extends TestCase
 
     public function testResourceAuthorizerIsNotResourceAuthorizer()
     {
-        $mock = $this->createMock(AuthorizerInterface::class);
+        $mock = $this->createMock(AdapterInterface::class);
         $this->resolver->method('getAuthorizerByResourceType')->willReturn(get_class($mock));
         $this->illuminateContainer->method('make')->willReturn($mock);