+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file. This projec
 ## Unreleased
 
 ### Added
+- New authorizer interface and an abstract class that better integrates with Laravel's authentication and
+authorization style. See the new [Security chapter](./docs/basics/security.md) for details.
 - The JSON API controller now has the following additional hooks:
   - `searching` for an *index* action.
   - `reading` for a *read* action.